requirements.txt
uvicorn
fastapi
deepgram-sdk
python-dotenv
httpx
litellm